zfb_tools/requirements.txt
aliyun-iot-linkkit==1.2.8
aliyun-python-sdk-core==2.13.26
aliyun-python-sdk-ecs==4.19.12
aliyun-python-sdk-iot==8.52.0
aliyun-python-sdk-kms==2.16.0
aliyun-python-sdk-rds==2.5.1
aliyun-python-sdk-slb==3.2.18
altgraph==0.17.3
certifi==2021.5.30
cffi==1.15.1
charset-normalizer==2.0.12
crcmod==1.7
cryptography==38.0.4
dataclasses==0.8
et-xmlfile==1.1.0
future==0.18.2
idna==3.4
importlib-metadata==4.8.3
importlib-resources==5.4.0
jmespath==0.10.0
libusb==1.0.24b1
lxml==4.9.1
MouseInfo==0.1.3
Naked==0.1.32
openpyxl==3.1.1
oss2==2.16.0
packaging==21.3
paho-mqtt==1.6.1
pefile==2022.5.30
Pillow==8.4.0
pkg-about==1.0.3
psutil==5.9.4
pyaliyunsdk==0.0.12
PyAutoGUI==0.9.54
pycparser==2.21
pycryptodome==3.16.0
pygame==2.4.0
PyGetWindow==0.0.9
pyinstaller==4.10
pyinstaller-hooks-contrib==2022.0
PyMsgBox==1.0.9
PyMySQL==1.0.2
pyOpenSSL==23.2.0
pyparsing==3.1.1
pyperclip==1.8.2
PyQt5==5.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.1
PyRect==0.2.0
PyScreeze==0.1.30
pyserial==3.5
pytweening==1.1.0
pyusb==1.2.1
PyVISA==1.11.3
pywin32==305
pywin32-ctypes==0.2.0
pywinusb==0.4.2
PyYAML==6.0
requests==2.27.1
shellescape==3.8.1
sip==4.19.8
six==1.16.0
typing_extensions==4.1.1
urllib3==1.26.13
usb==0.0.83.dev0
wincertstore==0.2
zipp==3.6.0
